Refactor legoomni into legoomni/misc
