+++ LEGO1/lego/legoomni/src/infocenter/elevatorbottom.cpp
@@ -5,8 +5,8 @@
 #include "legocontrolmanager.h"
 #include "legogamestate.h"
 #include "legoinputmanager.h"
-#include "legoomni.h"
 #include "legovariables.h"
+#include "misc.h"
 #include "mxmisc.h"
 #include "mxnotificationmanager.h"
 #include "mxtransitionmanager.h"
