+++ LEGO1/lego/legoomni/src/infocenter/infocenter.cpp
@@ -11,10 +11,10 @@
 #include "legocontrolmanager.h"
 #include "legogamestate.h"
 #include "legoinputmanager.h"
-#include "legoomni.h"
 #include "legoplantmanager.h"
 #include "legoutil.h"
 #include "legovideomanager.h"
+#include "misc.h"
 #include "mxactionnotificationparam.h"
 #include "mxbackgroundaudiomanager.h"
 #include "mxcontrolpresenter.h"
