+++ LEGO1/lego/legoomni/src/police/policestate.cpp
@@ -2,6 +2,7 @@
 
 #include "islepathactor.h"
 #include "legoomni.h"
+#include "misc.h"
 #include "mxdsaction.h"
 #include "mxmisc.h"
 #include "police.h"
