+++ LEGO1/lego/legoomni/src/video/legomodelpresenter.cpp
@@ -5,9 +5,9 @@
 #include "legocharactermanager.h"
 #include "legoentity.h"
 #include "legoentitypresenter.h"
-#include "legoomni.h"
 #include "legovideomanager.h"
 #include "legoworld.h"
+#include "misc.h"
 #include "misc/legocontainer.h"
 #include "misc/legotexture.h"
 #include "misc/version.h"
