+++ LEGO1/lego/legoomni/src/video/mxtransitionmanager.cpp
@@ -4,6 +4,7 @@
 #include "legoutil.h"
 #include "legovideomanager.h"
 #include "legoworld.h"
+#include "misc.h"
 #include "mxbackgroundaudiomanager.h"
 #include "mxmisc.h"
 #include "mxparam.h"
